MSQL-15: Polish before linking webservice and actual implementation. Prepares code to remove dependency on Quartz for scheduling.
git-svn-id: http://talendforge.org/svn/tom/trunk@89760 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.core/src/com/amalto/core/storage/task/staging/ExecutionStatistics.java
@@ -26,7 +26,7 @@ public class ExecutionStatistics {
 
     private String runningTime;
 
-    private int totalRecords;
+    private double totalRecords;
 
     private double processedRecords;
 
@@ -88,11 +88,11 @@ public class ExecutionStatistics {
     }
 
     @XmlElement(name = "total_record")
-    public int getTotalRecords() {
+    public double getTotalRecords() {
         return totalRecords;
     }
 
-    public void setTotalRecords(int totalRecords) {
+    public void setTotalRecords(double totalRecords) {
         this.totalRecords = totalRecords;
     }
